requirements.txt
psycopg2==2.9.5
django==4.1.7
Pillow==9.4.0
django-durationwidget==1.0.5
djangorestframework==3.14.0